treewide: Pass HTTP headers on through the track list
When playing media through QtMultimedia media-hub will enqueue it
through the track list first, adding the need for having HTTP headers
cached until use.

Since the track list is also part of the public API/MPRIS we can't just
extend it by changing the add_track method signature. Instead, since we
already have an OpenUriExtended DBus API not part of the MPRIS spec,
we set the headers right before the URI is added, so that the headers
are attached to their known track ID.