service: do not emit clientDisconnected signal on playbin reset
The playbin can get reset for many other reasons (such as missing
codecs) and the client must still be able to control the backend and
choose another file to play.

The clientDisconnected() signal is now unused, but let's keep it in case
we find real cases where the playbin can detect that the client
disconnected.