service: load media as soon as its URL is set
This is important when remote media is selected, because the MIME type
detection based solely on the URI can fail, whereas setting the
GStreamer pipeline to PAUSED state causes the media to start loading and
its channels are detected and counted.
Contextually, we also need to update the Engine class to not forward
these PAUSED events which are happening because of us performing this
trick.