tests: make track test more stable
The AboutToFinish signal is emitted 2-3 seconds before the actual
termination of the track. Since our default timeout for verifying
property changes is 3000ms, there's a concrete risk that our assertion
will fail.

Instead of just increasing the timeout in the property assertion, add
one more check to wait for the playback status to become "stopped",
which will happen when the track is effectively finished. We set this
timeout to 5000ms because, as said before, about 2 to 3 seconds will be
taken for effectively waiting for the song to finish, and another 2 can
be spent while D-Bus delivers the signal, if the machine is busy.