service: never request more than one lock at a time
The counters for the number of requested wakelocks needs to be in a
singleton class, the natural candidate being the PowerStateController
class, which is already a singleton and takes care of the states
requests.
We still need to keep a boolean flag in the PlayerImplementation, since
we are not very good at keeping track of the playback state. But once we
do, we could bind the wakelock requests to the playback state, and get
rid of the booleans.