service: update requested locks when video streams change
Make anotehr call to request_power_state() if the isVideoStream property
changes during playback. This ensures that we request the display lock,
if the media was initially detected as audio only (this typically
happens when playing remote video streams).

Calling request_power_state() multiple times is safe because it anyway
keeps tracks of the requested locks, and won't request them again.